pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "basicvsr-yuv420"
version = "0.1.0"
description = "BasicVSR training, evaluation, and inference package converted from the original notebook"
readme = "README.md"
requires-python = ">=3.7,<3.11"
dependencies = [
  "numpy>=1.21,<2",
  "Pillow>=9.5,<10",
  "torch==1.13.1",
  "torchmetrics>=0.11.4,<1.0",
  "tqdm>=4.64",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
